Refactor
- Remove ImageMagick dependency
- Add libjpeg-turbo dependency

This code change switches from ImageMagick to libjpeg-turbo.

libjpeg-turbo is available on most Linux OSes (verified:
ArchLinux, Ubuntu, Debian) and the version of ImageMagick differs
a lot. This is a problem because Version 6 and 7 of that library are not
source code compatible and it seems the code in use surfaces a bug where
the image data is correctly decoded whereas the metadata (size!) is not.

By removing ImageMagick completely and use the jpeg library directly
we can work around that problem. Sideeffect: If a camera preview does
not in fact return a JPEG image the code does not even try to decode the
image. (Not a problem for Canon cameras iff you set the capture mode to
JPEG and not RAW!)